backend/requirements.txt
fastapi
uvicorn
PyPDF2
groq
requests
pandas
json-repair
python-dotenv
